% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVar.R
\name{CVar}
\alias{CVar}
\title{calculate the coefficient of variation}
\usage{
CVar(x, remove_na = FALSE, ...)
}
\arguments{
\item{x}{a numeric vector.}

\item{remove_na}{a logical value indicating whether NA values should be contained in the computation. The default value is \emph{FALSE}.}

\item{...}{further arguments passed to or from other methods}
}
\value{
A vector containing the coefficient of variation of the input data
}
\description{
calculate the coefficient of variation
}
\examples{
CVar(c(1:4))
}
