% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_mean.R
\name{group_mean}
\alias{group_mean}
\title{A group_by mean Function}
\usage{
group_mean(data, group, value)
}
\arguments{
\item{data}{data you are working with}

\item{group}{variable you want to tab}

\item{value}{variable you want to tab}
}
\description{
This function allows you to get means of a each variable group
}
\examples{
group_mean()

}
