% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range.R
\name{range}
\alias{range}
\title{calculate the range}
\usage{
range(x, remove_na = FALSE)
}
\arguments{
\item{x}{a numeric vector.}

\item{remove_na}{a logical value indicating whether NA values should be contained in the computation. The default value is \emph{FALSE}.}
}
\value{
A vector containing the range of the input data
}
\description{
calculate the range
}
\examples{
range(c(1,5,9,0,NA),remove_na = TRUE)
}
