% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R
\name{tab}
\alias{tab}
\title{A Tabulate Function}
\usage{
tab(dataset, var)
}
\arguments{
\item{dataset}{data you are working with}

\item{var}{variable you want to tab}
}
\description{
This function allows you to create one-way tabulations (two-way coming soon!).
}
\examples{
tab()
}
\keyword{tab}
